Moved code from FeatureExtractor to Tab
Made large changes to the code base in order to allow easy reuse of methods when creating combined arff files
