+++ TabDifficultyAnalyser/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Owner\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=/Users/imac5/Library/Application Support/NetBeans/8.2/build.properties
